Porting tern analysis of JS code to Node. (#11948)
* Porting tern analysis of JS code to Node.

* Removing old tern submodule config

* JS text cache in HTML Mixed Mode.
gitmodule mapping changes.

* Submodules fix

* Fixing more submodule issues

* Tern default defs path fix

* First draft for content extraction in node [UNIT TESTING IN PROGRESS]

* Update dirty doc content in tern cache.

* Stability fix.

* Update submodule repo link.
Adding JSDoc in ExtractFileContent

* Submodule fix

* build error fix

* Usage of global use strict

* Node dependencies

* Remove redundant module defs

* Remove redundant script

* src/extensions/default/JavaScriptCodeHints/node/node_modules/acorn

* Update .gitmodules

To make it even with master

* Addressed review comments

* removed redundant for loop(Already added map function for the same)

* Fix for analysis data reset on project change

* Bug fixes for Untitled document mode change to 'javascript'

* Remove tern web worker

* Update Acorn and Tern to latest available version

* Bug fixes - Handle require plugin, untitled doc, doc comment plugin
