+++ src/extensions/default/JavaScriptCodeHints/MessageIds.js
@@ -35,7 +35,9 @@ define(function (require, exports, module) {
         TERN_GET_GUESSES_MSG        = "GetGuesses",
         TERN_WORKER_READY           = "WorkerReady",
         TERN_INFERENCE_TIMEDOUT     = "InferenceTimedOut",
-        SET_CONFIG                  = "SetConfig";
+        SET_CONFIG                  = "SetConfig",
+        TERN_UPDATE_DIRTY_FILE      = "UpdateDirtyFileEntry",
+        TERN_CLEAR_DIRTY_FILES_LIST = "ClearDirtyFilesList";
 
     // Message parameter constants
     var TERN_FILE_INFO_TYPE_PART    = "part",
@@ -58,6 +60,8 @@ define(function (require, exports, module) {
     exports.TERN_FILE_INFO_TYPE_EMPTY   = TERN_FILE_INFO_TYPE_EMPTY;
     exports.TERN_INFERENCE_TIMEDOUT     = TERN_INFERENCE_TIMEDOUT;
     exports.SET_CONFIG                  = SET_CONFIG;
+    exports.TERN_UPDATE_DIRTY_FILE      = TERN_UPDATE_DIRTY_FILE;
+    exports.TERN_CLEAR_DIRTY_FILES_LIST = TERN_CLEAR_DIRTY_FILES_LIST;
 });
