+++ src/extensions/default/JavaScriptCodeHints/node/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "brackets-javascript-code-hints",
+  "dependencies": {
+    "acorn": "5.1.1",
+    "tern": "0.21.0"
+  }
+}
